Move tests to device-type specific test files to make improvements more approachable
